+++ src/main/kotlin/lean4ij/language/Lean4SyntaxHighlighter.kt
@@ -19,7 +19,7 @@ import com.intellij.psi.PsiElement
 import com.intellij.psi.TokenType.WHITE_SPACE
 import com.intellij.psi.tree.IElementType
 import com.intellij.psi.util.elementType
-import lean4ij.Lean4Settings
+import lean4ij.setting.Lean4Settings
 import lean4ij.language.psi.TokenType
 import java.nio.charset.StandardCharsets